Finished pretty much my entire part
Added file IO to museum, enterprise, events.
Added google tests because Cucx exists
"Fixed" bug with Address, as when constructed with Region only constructor, it would write an invalid Address which would break stuff.
Moved a couple inline operator== from Event which were defined in the .cpp (should be in the header)