Created StateWorker class. Fix minor things in Card class
StateWorker class is and its hash table is done: Already defined its hashtable operators and created a type for it (StateWorkerHashTable). Need to make the museum netowkr write its workers to files, read them, list them (as we did for the cards), update their info (setters are done) and hire/fire them (functions for that are implemented): no need to delete them because their records are supposed to be permanent.